Convenience functions to set, rename, or validate a repo's remotes

The function add_bioc_remote adds an 'upstream' remote that points to git@git.bioconductor.org. If an upstream remote already exists, it will be validated against the SSH URL. By design, the Bioconductor remote location is called the 'upstream' remote. The 'origin' is set to the GitHub location.

add_bioc_remote(repo = ".")

set_bioc_remotes(repo = ".", org = "Bioconductor")

Arguments

repo

character(1) The local path to the git repository whose upstream remote should be set

org

character(1) The organization for which to extract the names of the repositories on GitHub (default "Bioconductor").

remote

character(1) The name of the remote to be created. This is usually named 'upstream' (default)

Value

add_bioc_remote: adds an 'upstream' remote with the Bioconductor git address for a given package.

set_bioc_remotes: updates the remotes for a given package, setting the 'origin' remote to the GitHub location and the 'upstream' remote to the Bioconductor git server.

Details

The set_bioc_remotes function will update a repository's remotes by setting origin to the GitHub location and upstream to the Bioconductor git server. If the origin remote is not set to the GitHub location, the function will overwrite the URL. If the upstream remote is not set, the function will add it. If the upstream remote is set, the function will overwrite the URL with the Bioconductor SSH URL. Note that the remotes will follow this template:

origin git@github.com:{org}/{package}
upstream git@git.bioconductor.org:packages/{package}