Convenience function to create the devel branch for all GitHub packages

This function identifies an organization's repositories that are packages given the current version of Bioconductor (from BiocManager::version()) and identifies which repositories need to have a devel branch added. It then adds the devel branch using the rename_branch_to_devel function. It is highly recommended that the user run this on the devel version of Bioconductor to avoid missing packages that are only in devel.

branch_all_packages(
  packages = character(0L),
  version = BiocManager::version(),
  old_branches = c(.OLD_DEFAULT_BRANCH, "main"),
  org = "Bioconductor",
  set_upstream = c("origin/devel", "upstream/devel"),
  clone = TRUE
)

Arguments

packages

named character() A character vector of default branches whose names correspond to Bioconductor package names. See packages_with_default_branch.

version

character(1) The current development version of Bioconductor given by BiocManager::version() (default). It is used to obtain the packages currently available in the Bioconductor CRAN-like repository with BiocManager::repositories().

old_branches

character() A vector of default branch names to be replaced, by default this includes 'master' and 'main'

org

character(1L) The organization for which to extract the names of the repositories on GitHub (default "Bioconductor").

set_upstream

character(1) The remote location that will be tracked by the local branch, either "origin/devel" (default) or "upstream/devel"

clone

logical(1) Whether to clone the GitHub repository into the current working directory (default: TRUE)

Details

Note that the clone argument allows the user to clone the repository first from GitHub via SSH. It is recommended that this be enabled and that the user running this function can clone packages via SSH and have access to modifying packages on the GitHub organization.

See also