Compare the old and current releases to generate the release announcements NEWS compilation

The function uses previous and current versions of Bioconductor to generate a single package's NEWS file.

getPackagesNEWS(
  prevRepos = "3.15",
  currRepos = "3.16",
  repo = c("bioc", "data/experiment", "workflows"),
  srcdir = NULL
)

printNEWS(
  dbs,
  destfile,
  overwrite = FALSE,
  width = 68,
  output = c("md", "text"),
  relativeLink = FALSE,
  ...
)

Arguments

prevRepos

character(1) The version string indicating the old release version of Bioconductor

currRepos

character(1) The version string indicating the newest and current release version of Bioconductor

repo

character(1) The repository nickname indicating which repository to compare news

srcdir

(Optional) character(1) The source directory in which all the Bioconductor packages, whose NEWS files are to be interrogated, reside

dbs

list() The news_db list as returned by getPackageNEWS.

destfile

character(1) The file path location to where NEWS will be saved.

overwrite

logical(1) Indicate whether destfile should be over-written, if it exists.

width

numeric(1) The maximum number of characters NEWS items can span before wrapping, excludes indent.

output

character(1) Either 'md' or 'text' indicating the output in markdown or text format, respectively.

relativeLink

logical(1) Whether to include links relative to the Bioconductor website or provide the full URL, i.e., https://bioconductor.org (default: FALSE).

...

additional arguments, unused.

Value

A list of NEWS