Generate the list of packages to be updated

This function obtains all the repositories from the Bioconductor organization and filters them to only valid R packages and repositories that do not have a RELEASE_XX_YY branch.

packages_list_to_be_updated(
  version = "3.16",
  org = "Bioconductor",
  type = c("BioCsoft", "BioCexp")
)

Arguments

version

character(1L) The numeric version of the Bioconductor release, e.g., "3.16"

org

character(1L) The organization for which to extract the names of the repositories on GitHub (default "Bioconductor").

type

character(1L) The official repository name as given by BiocManager::repositories(). Currently, only software and experiment data ('BioCsoft' and 'BioCexp', respectively) are supported.

Value

A named scalar string of the default branch whose name corresponds to a Bioconductor GitHub repository