Generate the list of packages to be updated

These functions obtain all the repositories from the designated organization and filters them to only valid R packages and repositories that do not have a RELEASE_X_Y branch or that do, depending on the function called.

packages_without_release_branch(
  version = "3.16",
  org = "Bioconductor",
  type = c("BioCsoft", "BioCexp")
)

packages_with_release_branch(
  version = "3.16",
  org = "Bioconductor",
  type = c("BioCsoft", "BioCexp")
)

Arguments

version

character(1) The numeric version of the Bioconductor release, e.g., "3.16"

org

character(1) The organization for which to extract the names of the repositories on GitHub (default "Bioconductor").

type

character() The repository names to look through as returned by BiocManager::repositories(). Currently, only software and experiment data ('BioCsoft' and 'BioCexp', respectively) are supported.

Value

A named scalar string of the default branch whose name corresponds to a Bioconductor GitHub repository

Examples

if (interactive()) {
    packages_without_release_branch(version = "3.19")
}