Create the 'devel' branch locally and on GitHub

The function is meant to be run one level up from the local git repository. It will create the 'devel' branch and push to the origin remote which should be set to GitHub. Upstream tracking can be configured to either the origin or upstream remote.

rename_branch_to_devel(
  package_name,
  from_branch = .OLD_DEFAULT_BRANCH,
  org = "Bioconductor",
  set_upstream = c("origin/devel", "upstream/devel"),
  clone = FALSE,
  is_bioc_pkg = TRUE
)

Arguments

package_name

character(1) The name of the organization R package that is also available on GitHub.

from_branch

character(1) The old default branch from which to base the new 'devel' branch from (default: 'master')

org

character(1) The organization for which to extract the names of the repositories on GitHub (default "Bioconductor").

set_upstream

character(1) The remote location that will be tracked by the local branch, either "origin/devel" (default) or "upstream/devel"

clone

logical(1) Whether to clone the GitHub repository into the current working directory (default: TRUE)

is_bioc_pkg

logical(1) Whether the repository is an R package that has an upstream remote on Bioconductor, i.e., <git@git.bioconductor.org:packages/Package>. If so, additional validity checks will be run on the git remotes.

Value

Called for the side effect of creating a 'devel' branch on the local and remote repositories on GitHub

Details

The origin remote is assumed to be GitHub, i.e., git@github.com:user/package but this requirement is not checked or enforced; thus, allowing flexibility in remote origin locations. The upstream remote is validated against the Bioconductor git repository address, i.e., git@git.bioconductor.org:packages/package. The local repository is validated before the devel branch is created.

Examples

if (interactive()) {

  rename_branch_to_devel(
    package_name = "SummarizedExperiment",
    org = "Bioconductor",
    set_upstream = "upstream/devel"
  )

}